+++ ConsoleGame/MainCircle.h
@@ -10,7 +10,7 @@ public:
 
 private:
 	clock_t  previous_time;
-	int frame_rate;
+	float frame_rate;
 	clock_t time_between_frames;
 };
